requirements.txt
dnspython==2.7.0
email_validator==2.2.0
idna==3.10
numpy==2.3.2
pandas==2.3.1
pyarrow==21.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
tzdata==2025.2
